% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist_openmp}
\alias{pdist_openmp}
\alias{pdist_c}
\alias{pdist_r}
\title{Calculate distance from origin of each point.}
\usage{
pdist_openmp(x, y, nthreads = 2)

pdist_c(x, y)

pdist_r(x, y)
}
\arguments{
\item{x, y}{numeric vectors of the same length}

\item{nthreads}{number of threads to run. Default: 2. Note: this 
number will only have an effect if code is actually compiled
with OpenMP support.  See \code{\link{has_openmp}()}}
}
\value{
distance from origin of each point
}
\description{
Calculation is done in parallel with OpenMP if available.
}
\details{
\describe{
  \item{\code{pdist_r}}{performs the calculation using R}
  \item{\code{pdist_c}}{performs the calculation using vanilla serial C code}
  \item{\code{pdist_openmp}}{performs the calculation using parallised code
   if OpenMP is available, otherwise uses serial C code}
}
}
\examples{
x <- c(1, 2)
y <- c(3, 4)
sqrt(x*x + y*y)
pdist_r(x, y)
pdist_c(x, y)
pdist_openmp(x, y)
}
