% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleep_c}
\alias{sleep_c}
\alias{sleep_openmp}
\title{Sleep inside a forloop in C in serial or in parallel}
\usage{
sleep_c(N = 2)

sleep_openmp(N = 2, nthreads = 2)
}
\arguments{
\item{N}{number of 1 second sleeps}

\item{nthreads}{number of threads}
}
\value{
None
}
\description{
Sleep inside a forloop in C in serial or in parallel
}
\examples{
system.time(sleep_c(2))
system.time(sleep_openmp(2, nthreads = 3))
}
